requirements.txt
amqp==2.3.2
billiard==3.5.0.4
certifi==2018.10.15
chardet==3.0.4
idna==2.7
kombu==4.2.1
Pillow==5.3.0
pip==18.1
pytz==2018.7
requests==2.20.1
setuptools==40.6.2
urllib3==1.24.1
vine==1.1.4
celery==4.2.1
Django==1.11.18
djangorestframework==3.8.2
scikit-learn==0.19.1
torch==0.3.1
opencv-python